requirements.txt
dialogflow~=1.1.1
python-telegram-bot~=13.7
python-dotenv~=0.19.0
requests~=2.26.0
telegram~=0.0.1
vk-api~=11.9.4